move most of the data manipulation to a new file
